Use `forEach` instead of `for (var x in foo)`.
Given the supported browsers, it is safe to use `forEach` which is a
better version of `for (var x in foo)`.

Since `for` iterates on the `enumerable properties`, it can lead to
unexpected behaviors when you are extending the array prototype. As an
example, check the following ember-twiddle, which show all the other
returned properties when using `for x in foo` in Ember.

https://ember-twiddle.com/bac28d42c800c4dffc4253f66afd0cac?openFiles=controllers.application.js%2C.