Fixed cast and session state handling
Plyr cast button now responds properly to current session state.
The button now represents whether the parent plyr is the one
that is currently assigned to ongoing cast session.

On page load, if there is no cast session, clicking the button
does two things:
    - Show cast drop down
    - Bind current plyr instance on successful session negotiation

However, on page load, if there already exists a cast session, the
button is still greyed out to signal that the plyr is not attached.
Upon click, the plyr is bound to the cast session and only the
icon is updated without going through cast connection.