Update types.js
Removing the providers constant, it is now a static property on Plyr class. This way it can be augmented through the static "use" method and can be accessed through all Plyr class instanciation. The providers static property is now an array of PlyrProvider class.

getProviderByUrl is also moved as a static method inside Plyr and returns a PlyrProvider class instead of a string.