Update defaults.js
Config for both vimeo and youtube has been moved to their provider class. Every PyrProvider has a static config property. When adding the PyrProvider to Pyr, the config is set inside config[PyrProvider.name].

The only YouTube event still relevant is statechange (qualitychange support has been removed in a previous release). This event is moved to the static events method of the YouTubeProvider. Each PlyrProvider has a static events method returning an array of additional event specific to the provider.