Update media.js
Pointing to the new way to get the name of the provider for the css class.

Removing the conditions to better implement new external providers. Since the provider is set before media.setup is called, there is no reason why not just call this.provider.setup.

Also moving specific configuration of speed (youtube.js and vimeo.js) here. So the speed availability is always automatically set according to the provider speed available and the options configured.