Update plyr.js
At every new instance, each provider specific config and events are added to it.

getProviderByUrl is now a static method of Plyr.

this.config.playsinline is true by default and is now modified in the beforeSetup static method for YouTube.

New getProvider static method to get the PyrProvider class by its name.

As long as now provider is found, this.provider is undefined.

Type depend on what the provider decides it to be. Some external provider might be of Audio types.

HTML5Provider can be of Video or Audio type, therefore the static method decides it. This is also an example for provider that have both types.

When destroying a player, the static method destroy is called on the provider asyncally. When it's done, the callback is called to finish destroying.