Set PIP support based on the Picture in Picture API
Previously, picture in picture support was explicitly disabled on iPhones,
and used the webkit specific APIs for Safari in general.

Safari (both mobile and desktop) has supported the Picture in Picture API
since September 15, 2020 (https://caniuse.com/picture-in-picture) so there's
no need to use the webkit API.

Additionally, picture in picture support appears to work on iPhones with this
API, so the check is no longer necessary.

Tested on:

- iPhone 14 (physical device)
- iPad 9th gen (browserstack.com)

Negative tests

- iPhone 11 (browserstack.com) - predates support but does not raise errorsoes not raise errors